We got some boilerplate in place
Ok so we refactored the code so we have some boilerplate in the main
function, and we have some tests, and we have some of the structure such
that we can start working on the project. What we need to do is to
actually be able to start integrating each of these pieces. I think we
start with the connectors, we make sure the docker-compse spins up each
of the server types with dummy info, and the unit test start testing out
these connectors.