Renaming the project and refactoring
Now that we have a new name, maybe it is a good opertunity to do some
refactoring. But instead we have made more of the unreadable, unrefined
mess of a project. So far we have way too many folders, and files, there
is no reason why we need to be so modular, we have taken object oriented
programming to unnacisarily complex levels. Maybe we ditch all the
excessive files and structures come up with a simple folder startucture
and move forward